long doubles

+++ codes/referee_questions/prepare_mock/inc/config.h
@@ -11,18 +11,18 @@
 
 /* Different coordinate data of a given star */
 typedef struct STAR{
-    double gal_z, gal_r, gal_phi; // gal-centered gal coord.
-    double gal_l_rad, gal_b_rad; // sun-centered gal coord.
-    double ra_rad, dec_rad, distance; // sun-centered equatorial coord.
-    double x, y, z; // sun-centered (equatorial/cartesian)
+    long double gal_z, gal_r, gal_phi; // gal-centered gal coord.
+    long double gal_l_rad, gal_b_rad; // sun-centered gal coord.
+    long double ra_rad, dec_rad, distance; // sun-centered equatorial coord.
+    long double x, y, z; // sun-centered (equatorial/cartesian)
 } STAR;
 
 /* Conversions between coordinates systems */
-double get_distance(double Z, double R, double phi);
+long double get_distance(long double Z, long double R, long double phi);
 void ZR_to_gal(STAR *s);
 void gal_to_eq(STAR *s);
 void eq_to_cart(STAR *s);
 
 /* Generation of random values */
-double random_gal_Z(double z0, double pdf_norm, double z_min, double z_max);
-double random_gal_R(double r0, double pdf_norm, double r_min, double r_max);
+long double random_gal_Z(long double z0, long double pdf_norm, long double z_min, long double z_max);
+long double random_gal_R(long double r0, long double pdf_norm, long double r_min, long double r_max);
